# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Fem-ho
  version: 0.1.0
  description: |
    Contracte de l'API de Fem-ho. Aquest fitxer és la font de veritat (regla 5
    d'instruccions.md): els tipus de TypeScript i el client de Kotlin es generen des
    d'aquí i cap client escriu tipus a mà. Un endpoint que no hi és no existeix.

    A M1 només hi ha les dues rutes públiques de vida i identitat. La resta arriba a
    la seva fita, seguint docs/05-api-rest.md.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: /api/v1
    description: Base de l'API. GET /info i GET /healthz pengen de l'arrel, no d'aquí.
...
paths:
...
  /tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: Tasques amb filtres
      parameters:
        - name: scope_id
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          description: Valors múltiples separats per comes.
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: cursor
          in: query
          description: Cursor opac. El client no l'ha d'interpretar mai.
          schema:
            type: string
      responses:
        '200':
          description: Pàgina de tasques.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    post:
      tags: [tasks]
      operationId: createTask
      summary: Crear una tasca
      description: |
        **`scope_id` és obligatori.** És la invariant central del producte: pot ser que
        un àmbit tingui una tasca sense projecte, però mai una tasca sense àmbit
        (docs/01 §4). L'`id` i la `position` els pot posar el client (D3, D4).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '200':
          description: Ja existia amb aquest `id`. Idempotència.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Falta l'àmbit, o no és vàlid.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /events:
    get:
      tags: [events]
      operationId: listEvents
      summary: Esdeveniments en una finestra
      description: |
        **Requereix `from` i `to`**: sense finestra no es poden expandir repeticions
        (docs/05 §4). Una RRULE sense COUNT ni UNTIL és infinita.

        Torna OCURRÈNCIES, no files: una sèrie setmanal dins d'un mes surt quatre
        vegades, cadascuna amb el seu `recurrence_id`.

        **Els esdeveniments no surten mai al kanban** (D8). Aquesta és l'única manera
        de demanar-los.
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: scope_ids
          in: query
          description: Àmbits actius, separats per comes.
          schema: { type: string }
      responses:
        '200':
          description: Ocurrències dins de la finestra, ordenades per inici.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventOccurrence'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          description: Falta la finestra, o no és vàlida.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [events]
      operationId: createEvent
      summary: Crear un esdeveniment
      description: |
        Un esdeveniment **no és una tasca amb hores** (D8): té els seus propis `status`
        —TENTATIVE, CONFIRMED, CANCELLED—, té `transparency`, i un calendari extern
        subscrit només en pot produir d'aquests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInput'
      responses:
        '201':
          description: Creat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '200':
          description: Ja existia amb aquest `id`. Idempotència.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /tasks/{id}/checklists:
    get:
      tags: [checklists]
      operationId: listChecklists
      summary: Les llistes senzilles d'una tasca
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Llistes amb els seus ítems.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Checklist'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [checklists]
      operationId: createChecklist
      summary: Crear una llista dins d'una tasca
      description: |
        Una llista pertany **sempre** a una tasca, i opcionalment s'ancora a una subtasca
        concreta (P1). No és un subprojecte ni una subtasca: són dues taules i un flag de
        presentació.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistInput'
      responses:
        '201':
          description: Creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checklist'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /checklists/{id}/items:
    post:
      tags: [checklists]
      operationId: createChecklistItem
      summary: Afegir un ítem
      description: |
        Un ítem **només té text i fet/no fet** (P1). Cap data, cap assignat, cap
        niuament. La contenció és deliberada i ve de Things 3: és el que fa que la llista
        no es converteixi en un segon gestor de tasques dins del gestor de tasques.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistItemInput'
      responses:
        '201':
          description: Afegit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItem'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags: [checklists]
      operationId: listChecklistItems
      summary: Els ítems d'una llista
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Els ítems.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ChecklistItem' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /sync:
    get:
      tags: [sync]
      operationId: pullSync
      summary: Delta des d'un cursor
      description: |
        Ordenat per `seq` ascendent, sempre. Filtrat pel principal: només arriba el dels
        àmbits que el token pot veure.

        Un `upsert` porta l'entitat SENCERA, no un diff: és més trànsit i molta menys
        complexitat, i les entitats de Fem-ho són petites. Un `delete` només porta
        l'identificador.

        **Sense cursor és una sincronització completa.**
      parameters:
        - name: cursor
          in: query
          description: Cadena OPACA. El client no l'ha d'interpretar mai.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 500 }
      responses:
        '200':
          description: El delta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '409':
          description: |
            El cursor és més vell que la retenció de tombstones i el delta seria
            incomplet. El client buida la base local i torna a començar sense cursor.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /sync/batch:
    post:
      tags: [sync]
      operationId: pushSync
      summary: Buidat de la cua de sortida
      description: |
        **Cada operació es resol per separat**: una que falli no ha de tombar el lot.

        `op_id` és la clau d'idempotència. Reenviar un lot després d'una caiguda no
        duplica res.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncBatch'
      responses:
        '200':
          description: Un resultat per operació, en el mateix ordre.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBatchResults'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /tasks/{id}/subtasks:
    get:
      tags: [tasks]
      operationId: listSubtasks
      summary: Les subtasques d'una tasca
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Les subtasques.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Subtask' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [tasks]
      operationId: createSubtask
      summary: Crear una subtasca
      description: |
        Una subtasca NO és una tasca: no té àmbit propi, ni estat de kanban, ni data, ni
        assignat. Té títol i fet/no fet, i pot ancorar una llista senzilla.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SubtaskInput' } } }
      responses:
        '201':
          description: Creada.
          content: { application/json: { schema: { $ref: '#/components/schemas/Subtask' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '422':
          description: Falta el títol.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
components:
  schemas:
...
    ChecklistInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id: { type: string }
        name: { type: string, minLength: 1 }
        subtask_id: { type: string }
        show_completed_inline: { type: boolean }
...
    Event:
      type: object
      additionalProperties: false
      required: [id, calendar_id, uid, summary, starts_at, all_day, version]
      properties:
        id: { type: string }
        calendar_id: { type: string }
        uid: { type: string, description: UID d'iCalendar. }
        recurrence_id:
          type: [string, 'null']
          description: Nul al mestre. Amb valor, és una instància modificada (D8).
        summary: { type: string }
        description: { type: [string, 'null'] }
        location: { type: [string, 'null'] }
        starts_at: { type: string, format: date-time }
        ends_at: { type: [string, 'null'], format: date-time }
        all_day:
          type: boolean
          description: |
            VALUE=DATE és tot el dia i NO té fus. Convertir-lo a mitjanit UTC és l'error
            que fa que els aniversaris apareguin el dia abans a mig món (docs/07 §8).
        timezone: { type: [string, 'null'] }
        status:
          type: string
          enum: [TENTATIVE, CONFIRMED, CANCELLED]
          description: Els de VEVENT. Els de VTODO són uns altres (D8).
        transparency:
          type: string
          enum: [OPAQUE, TRANSPARENT]
          description: Si ocupa temps. VTODO no en té equivalent.
        rrule: { type: [string, 'null'] }
        version: { type: integer }
    EventInput:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        calendar_id: { type: string }
        summary: { type: string, minLength: 1 }
        description: { type: string }
        location: { type: string }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        all_day: { type: boolean }
        timezone: { type: string }
        status:
          type: string
          enum: [TENTATIVE, CONFIRMED, CANCELLED]
        transparency:
          type: string
          enum: [OPAQUE, TRANSPARENT]
        rrule: { type: string }
    EventOccurrence:
      type: object
      additionalProperties: false
      required: [event_id, uid, summary, starts_at, ends_at, all_day, scope_id]
      properties:
        event_id: { type: string }
        uid: { type: string }
        recurrence_id:
          type: [string, 'null']
          description: L'inici que li tocaria segons la regla, encara que s'hagi mogut.
        summary: { type: string }
        location: { type: [string, 'null'] }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        all_day: { type: boolean }
        scope_id: { type: string }
        calendar_id: { type: string }
        status:
          type: string
          enum: [TENTATIVE, CONFIRMED, CANCELLED]
        is_override:
          type: boolean
          description: Cert si aquesta ocurrència ve d'una instància modificada.
...
    Task:
      type: object
      additionalProperties: false
      required:
        [id, scope_id, title, status, position, ai_mode, view_mode, created_at, updated_at, version]
      properties:
        id: { type: string }
        scope_id: { type: string }
        project_id: { type: [string, 'null'] }
        title: { type: string }
        description: { type: [string, 'null'] }
        status:
          type: string
          enum: [inbox, todo, doing, done]
          description: La columna. Es diu `status`, no `column` (D2).
        position:
          type: string
          description: Índex fraccional. Ordenat amb collation binària (D3).
        due_date:
          type: [string, 'null']
          description: YYYY-MM-DD, sense fus. Una data de tot el dia no té instant.
        due_time: { type: [string, 'null'] }
        deadline: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        view_mode:
          type: string
          enum: [card, simple]
        ai_mode:
          type: string
          enum: [manual, assisted, delegated]
          description: Delegar no és assignar (D5).
        delegate_agent_id: { type: [string, 'null'] }
        rrule:
          type: [string, 'null']
          description: 'RFC 5545, o `null` si no es repeteix.'
        recurrence_mode:
          type: [string, 'null']
          enum: [schedule, completion, null]
          description: |
            `schedule` compta des del venciment; `completion`, des que es fa. És la
            distinció que Todoist escriu com a `every` contra `every!` i que RRULE no sap
            expressar (docs/01 §4).
        recurrence_parent_id:
          type: [string, 'null']
          description: La instància de la qual ve, si n'és una repetició.
        progress:
          type: object
          required: [done, total, lists]
          description: |
            Subtasques i ítems de llista **comptats junts**. La targeta els ensenya sota
            un sol commutador ("3/7 fets") perquè per a qui mira el tauler són el mateix:
            coses que falten dins d'aquesta tasca. La distinció importa en obrir-la.

            És un agregat i no la llista: portar tots els ítems de totes les targetes
            faria baixar milers de files per pintar unes quantes pastilles.
          properties:
            done: { type: integer }
            total: { type: integer }
            lists:
              type: integer
              description: |
                Quants blocs desplegables té la targeta: un per les subtasques, si en té,
                i un per cada llista. És el número del commutador "Llistes (2)", que
                compta blocs i no ítems.
        assignee_ids:
          type: array
          items: { type: string }
          description: Persones. El brief demana "persona o persones".
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer }
    TaskInput:
      type: object
      additionalProperties: false
      required: [scope_id, title]
      properties:
        id: { type: string }
        scope_id:
          type: string
          description: Obligatori. Mai una tasca sense àmbit (docs/01 §4).
        project_id: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        status:
          type: string
          enum: [inbox, todo, doing, done]
          default: inbox
        position: { type: string }
        due_date: { type: string }
        due_time: { type: string }
        assignee_ids:
          type: array
          items: { type: string }
...
    TaskPage:
      type: object
      additionalProperties: false
      required: [data, has_more]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        next_cursor: { type: [string, 'null'] }
        has_more: { type: boolean }
...
    Problem:
      type: object
      description: |
        Error en application/problem+json (RFC 9457, docs/05 §3).

        **Qui llegeix un error no és qui el pateix.** El reben tres públics molt
        diferents: una persona davant de l'app, un agent per MCP, i un client CalDAV com
        DAVx⁵ o Thunderbird. Els dos últims no tenen ni catàleg ni idioma. Per això
        l'error viatja en dues peces:

        - `type`, `title` i `detail` **en anglès**, estables, per programar-hi.
        - `params`, les **dades** de l'error. Les apps hi posen el text del seu catàleg
          amb la clau `error.<slug>`, on el `slug` és l'últim tram del `type`. Si no en
          tenen la clau, ensenyen el `detail`: així un error nou del servidor no deixa
          mai una pantalla muda ni obliga a desplegar les dues bandes alhora.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          examples: ['https://femho.app/errors/scope-forbidden']
        title:
          type: string
          examples: ['Scope not accessible']
        params:
          type: object
          additionalProperties: { type: [string, integer] }
          description: |
            Les dades de l'error, planes. Plans a posta: el client no ha de saber
            recórrer un JSON dins d'un JSON per pintar un missatge.
          examples: [{ name: 'Família', count: 3 }]
        status:
          type: integer
          examples: [403]
        detail:
          type: string
          examples: ["Aquest token només té accés a l'àmbit Feina."]
        instance:
          type: string
  responses:
    Unauthenticated:
      description: Cal autenticar-se.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: |
        Sense permís. El cos diu QUINS àmbits veu el token i ON és el que s'ha demanat:
        un 403 mut fa que un agent reintenti en bucle (docs/05 §2).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No existeix, o no és visible per a qui pregunta.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: femho_session
      description: Sessió de navegador. Galeta HttpOnly, Secure, SameSite=Lax.
    bearer:
      type: http
      scheme: bearer
      description: |
        Token d'accés (Android) o token d'API `femho_pat_…` (integracions, MCP, IA).
        Les capacitats i els àmbits van al registre del token, no a scopes d'OAuth
        (regla 9).
security:
  - session: []
  - bearer: []